% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packettotal-package.R
\docType{package}
\name{packettotal}
\alias{packettotal}
\alias{packettotal-package}
\title{...}
\description{
\itemize{
\item URL: \url{https://gitlab.com/hrbrmstr/packettotal}
\item BugReports: \url{https://gitlab.com/hrbrmstr/packettotal/issues}
}
}
\author{
Bob Rudis (bob@rud.is)
}
