% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api-key.R
\name{packettotal_api_key}
\alias{packettotal_api_key}
\title{Get or set PACKETTOTAL_API_KEY value}
\usage{
packettotal_api_key(force = FALSE)
}
\arguments{
\item{force}{Force setting a new PacketTotal key for the current environment?}
}
\value{
atomic character vector containing the PacketTotal api key
}
\description{
The API wrapper functions in this package all rely on a PacketTotal API
key residing in the environment variable \code{PACKETTOTAL_API_KEY}.
The easiest way to accomplish this is to set it
in the \code{.Renviron} file in your home directory.
}
\references{
\url{https://packettotal.com/api-docs/}
}
