% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep-search.R
\name{pt_deep_search}
\alias{pt_deep_search}
\alias{pt_get_search_results}
\title{Create a new deep search task. Search for a term or with a Lucene query.}
\usage{
pt_deep_search(query, api_key = packettotal_api_key())

pt_get_search_results(search_result, api_key = packettotal_api_key())
}
\arguments{
\item{query}{search term (e.g. an IP address, domain, or file hash) or valid Lucene query}

\item{api_key}{your \code{\link[=packettotal_api_key]{packettotal_api_key()}}.}

\item{search_result}{output from \code{\link[=pt_deep_search]{pt_deep_search()}} or a plain search results id}
}
\description{
Unlike the more lighweight \code{\link[=pt_search]{pt_search()}} results from this endpoint
will be available at the returned URL.
}
\examples{
str(try(pt_deep_search("botnet OR malware"), silent=TRUE), 1)
}
\references{
<https://packettotal.com/api-docs/#/search
}
