% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detail.R
\name{pt_detail}
\alias{pt_detail}
\title{Get a detailed report of PCAP traffic, carved files, signatures, and top-talkers.}
\usage{
pt_detail(pcap_id, api_key = packettotal_api_key())
}
\arguments{
\item{pcap_id}{An md5 hash corresponding to the PCAP file submission on PacketTotal.com.
This hash can be derived by hashing the PCAP file in question.}

\item{api_key}{your \code{\link[=packettotal_api_key]{packettotal_api_key()}}.}
}
\description{
Analysis results contain high-level protocol statistics, signatures, and intelligence that PacketTotal discovered during analysis and enrichment.
}
\examples{
str(try(pt_detail("d210f4dbea97949f694e849507951881"), silent=TRUE), 2)
}
\references{
\url{https://packettotal.com/api-docs/#/pcaps/get_pcaps}
}
