% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{pt_download}
\alias{pt_download}
\title{Download a PCAP analysis archive. The result is a zip archive containing the PCAP itself, CSVs representing various analysis results, and all carved files.'}
\usage{
pt_download(pcap_id, dl_dir = getwd(), archive_name = NULL,
  api_key = packettotal_api_key())
}
\arguments{
\item{pcap_id}{An md5 hash corresponding to the PCAP file submission on PacketTotal.com.
This hash can be derived by hashing the PCAP file in question.}

\item{dl_dir}{directory where to store the download}

\item{archive_name}{name of the ZIP file. If left \code{NULL} then a ZIP file
will be created with the name \code{YYYY-mm-dd-pcap_id.zip}.}

\item{api_key}{your \code{\link[=packettotal_api_key]{packettotal_api_key()}}.}
}
\value{
if successful and the analysis package is ready then the full path
to the ZIP file is returned (invisibly). If the analysis package
is not ready the return value is "\code{_PROCESSING_}".
}
\description{
Download a PCAP analysis archive. The result is a zip archive containing the PCAP itself, CSVs representing various analysis results, and all carved files.'
}
\examples{
str(try(pt_download("536cf06ca83704844d789f56caf22ee6"), silent=TRUE), 2)
}
\references{
\url{https://packettotal.com/api-docs/#/pcaps/get_pcaps__pcap_id__download}
}
