% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{pt_info}
\alias{pt_info}
\title{Get high-level information about a specific PCAP file.}
\usage{
pt_info(pcap_id, api_key = packettotal_api_key())
}
\arguments{
\item{pcap_id}{An md5 hash corresponding to the PCAP file submission on PacketTotal.com.
This hash can be derived by hashing the PCAP file in question.}

\item{api_key}{your \code{\link[=packettotal_api_key]{packettotal_api_key()}}.}
}
\description{
Results will contain high-level information, such as what logs were extracted, the date it was analyzed, and additional references.
}
\examples{
str(try(pt_info("d210f4dbea97949f694e849507951881"), silent=TRUE), 2)
}
\references{
\url{https://packettotal.com/api-docs/#/pcaps/get_pcaps}
}
