% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random.R
\name{pt_random}
\alias{pt_random}
\title{Get high-level information about a random PCAP file.}
\usage{
pt_random(api_key = packettotal_api_key())
}
\arguments{
\item{api_key}{your \code{\link[=packettotal_api_key]{packettotal_api_key()}}.}
}
\description{
Randomly selected PCAPs come from a set of pre-selected, interesting PCAP files.
}
\examples{
str(try(pt_random(), silent=TRUE), 1)
}
\references{
\url{https://packettotal.com/api-docs/#/pcaps/get_pcaps}
}
