% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{pt_search}
\alias{pt_search}
\title{Search with term or with a valid Lucene query.}
\usage{
pt_search(query, api_key = packettotal_api_key())
}
\arguments{
\item{query}{search term (e.g. an IP address, domain, or file hash) or valid Lucene query}

\item{api_key}{your \code{\link[=packettotal_api_key]{packettotal_api_key()}}.}
}
\description{
Receive a set of matches for given query.
}
\examples{
str(try(pt_search("evil.com"), silent=TRUE), 1)
}
\references{
<https://packettotal.com/api-docs/#/search
}
