% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph.R
\name{pt_similar}
\alias{pt_similar}
\title{Get a similarity graph relative to the current PCAP file.}
\usage{
pt_similar(pcap_id, weighting_mode = c("behavior", "content"),
  intensity = c("minimal", "low", "medium", "high"),
  prioritize_uncommon_fields = FALSE, api_key = packettotal_api_key())
}
\arguments{
\item{pcap_id}{An md5 hash corresponding to the PCAP file submission on PacketTotal.com.
This hash can be derived by hashing the PCAP file in question.}

\item{weighting_mode}{One of "\code{behavior}" (default) or "\code{content}". Weight search results either based on their similarity to the behaviors exhibited or contents contained within the current PCAP file.}

\item{intensity}{One of "\code{minimal}" (default), "\code{low}", "\code{medium}", or "\code{high}". The scope of the search, basically translates to the maximum number of aggregations to exhaust. Using a high level intensity, may result in occassional timeouts.}

\item{prioritize_uncommon_fields}{By default, the most common values are used to seed the initial similarity search. Enabling this parameter, seeds the initial search with the least common values instead.}

\item{api_key}{your \code{\link[=packettotal_api_key]{packettotal_api_key()}}.}
}
\description{
Results contain PCAPs that exhibit similar behaviors or contain similar content. Results are organized with the most similar PCAPs on top, and the terms that were found shared within both.
}
\examples{
str(try(pt_similar("536cf06ca83704844d789f56caf22ee6"), silent=TRUE), 3)
}
\references{
\url{https://packettotal.com/api-docs/#/pcaps/get_pcaps__pcap_id__similar}
}
