% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usage.R
\name{pt_usage}
\alias{pt_usage}
\title{Retrive usage and subscription plan information.}
\usage{
pt_usage(api_key = packettotal_api_key())
}
\arguments{
\item{api_key}{your \code{\link[=packettotal_api_key]{packettotal_api_key()}}.}
}
\description{
Handy helper to determine how many requests you have remaining.
}
\examples{
str(try(pt_usage(), silent=TRUE), 2)
}
